refactor(nika-event): the stamper seam descends — the wall, third payment
B6's wiring + B7's billing pushed nika-runtime back over its 15k prod
wall (15012). The descent this time: the stamper/sink family
(Stamper · EventSink · DeterministicStamper · SystemStamper · VecSink)
moves to nika-event — it IS event machinery (the journal's identity
and delivery seams), and the crate had the headroom. The runtime keeps
a path-stable shim (`stamp::` and the root re-exports are unchanged;
the public-api baseline shifts to the re-export form, source-compatible).

Also pays the file-loc-cap the same push crossed: cli-host's probe.rs
hit 1534 — the harness surface (registry→resolver rows · the combined
access-probe channel) lives in a new `harness.rs` beside it.
