+++ estate.yaml
@@ -18,11 +18,11 @@ classes:
   testimonial: "captured evidence — fixtures, corpora, traces, receipts: proves behavior; neither shipped source nor regenerable output"
   foreign: "pointer to a third-party sovereign source"
 summary:
-  classified_files: 4356
+  classified_files: 4358
   file_rows: 11
   pattern_rows: 22
   by_class:
-    authored: 1425
+    authored: 1427
     authored-pin: 2
     generated: 118
     pinned-copy: 115
@@ -140,7 +140,7 @@ patterns:
 - glob: "crates/*/public-api.txt"
   class: generated
   match_count: 62
-  aggregate_sha256: 6ed95ea77f14f069ea927749f582101e3528681e4b25a3b24d47e887fbc89fbb
+  aggregate_sha256: 02b3820ae3a093e95a5350ef8f3ebe20252b87bf2d8b59e92f6bd16f2303980e
   evidence: ".github/workflows/public-api.yml 'diff each covered lib crate' byte-compares each snapshot against cargo public-api output; macOS renders differently, so snapshots regenerate FROM THE UBUNTU ARTIFACT (workflow comment)"
   derivation:
     tool: "cargo public-api -p <crate> --omit auto-trait-impls > crates/<crate>/public-api.txt (DEFAULT features · the CI job shape — all-features drags platform deps into the render · cargo-public-api 0.51.0 pinned · the public-api-actual CI artifact stays the judge)"
@@ -151,18 +151,18 @@ patterns:
     - "cargo-public-api 0.51.0"
 - glob: "crates/**/src/**"
   class: authored
-  match_count: 720
-  aggregate_sha256: 42daf7a444efd4b77a2d3f8aa8f734a232c9ebe2d2a8f1a69ea6718a1bec9dc7
+  match_count: 722
+  aggregate_sha256: 88348e33a84d5bbe0123004c79a50687909113fcce75478ca204add8ead2a0ea
   evidence: "the code — SPDX AGPL-3.0-or-later headers · 12-gate admission (ADR-003) · zero-unwrap discipline; spot-checked file heads carry no generation marker (grep hits for 'generated' are prose mentions, not markers)"
 - glob: "crates/**/tests/**"
   class: authored
   match_count: 135
-  aggregate_sha256: 81efc4d260f7b356aa013e5d4b28a65e83e03117d609b1f2744aeb4eefb94e1a
+  aggregate_sha256: 1c99554215aa86ac15c1ef57d67378236503d06acb223c5cb51b661e74262133
   evidence: "crate-owned tests + fixtures under tests/ — written with the crate per gate discipline, no generation markers"
 - glob: "crates/**"
   class: authored
   match_count: 132
-  aggregate_sha256: 167f1c09142124f3424a3148b39d09fefd8e0f308b7e68061ba81c9d7c035a40
+  aggregate_sha256: b74233e3eda9951853cb69b57f1fa52d98e5c73531c0dc466c5656005529e0f3
   evidence: "the remaining crate surface — Cargo.toml manifests · build.rs · benches · READMEs · hand-curated data catalogs (llm-providers · mcp-servers · embeddings · model-capabilities, probed daily by catalog-verify.yml); the pricing snapshot is excepted in files:"
   note: "crates/nika-pack/scripts/sync-pack.sh is a DIVERGENT older duplicate of scripts/sync-pack.sh — the pack-resync.yml lane runs the root one"
 - glob: "fuzz/**"
